build_matrix.sh: Fix a shellcheck warning
The warning was:
Use "${var:?}" to ensure this never expands to /* . [SC2115]

+++ build_matrix.sh
@@ -58,7 +58,7 @@ for CC in ${MATRIX_CC:-gcc clang}; do
             echo 'Cleaning...'
             travis_fold start cleaning
             if [ "$CMAKE" = yes ]; then rm -rf build; else make distclean; fi
-            rm -rf "$PREFIX"/*
+            rm -rf "${PREFIX:?}"/*
             git status -suall
             # Cancel changes in configure
             git checkout configure
